Change the backdrop alpha in screen orientation or size changes.
This fixes #572. The main change is that `true` is passed as a
`forceLayout` parameter into `viewWillTransition(to:with:)` callbacks.

Because it's necessary for the backdrop alpha's update when the view
size or its size classes changes.

This also fixes a regression by 9c45c31.

```diff
-        layoutAdapter.activateLayout(for: state, forceLayout: true)
+        layoutAdapter.activateLayout(for: state, forceLayout: forceLayout)
```

The behavior before the above second change indicates that the method is
working well even when `forceLayout` is set to `true` in their callbacks.