Display the scroll indicator of table view in Maps example
The top inset of its scroll indicator has an unexpected top margin.
`UITableView` seems to append a top inset in addition to its scroll
insets. However, it's important to display the indicator for the library
testing. This allows us to verify whether the scroll indicator shows and
hides as expected.