Allow delegate to decide whether the scroll view should be locked during state transitions
In #587 the behavior of the panel was changed so that the content of the tracked scroll view is always locked during state transitions. In some cases it's useful to be able to adjust the scroll offset during these transitions, so a delegate method has been introduced letting the user of the library decide on this behavior.